--- app/src/main/java/net/svishch/android/githubclient/mvp/model/CountersModel.kt
@@ -1,15 +1,0 @@
-package net.svishch.android.githubclient.mvp.model
-
-class CountersModel {
-    private val counters = mutableListOf ( 0 , 0 , 0 )
-    private fun getCurrent (index: Int ) : Int {
-        return counters [ index ]
-    }
-    fun next ( index: Int ) : Int {
-        counters [ index ] ++
-        return getCurrent ( index )
-    }
-    fun set ( index: Int , value: Int ){
-        counters [ index ] = value
-    }
-}
