+++ app/src/main/java/net/svishch/android/githubclient/mvp/model/Model.kt
@@ -1,7 +1,15 @@
 package net.svishch.android.githubclient.mvp.model
 
-import java.util.*
-
 class Model {
-    private val counters: MutableList<Int> = ArrayList(Arrays.asList(0, 0, 0))
+    private val counters = mutableListOf (0,0,0)
+    private fun getCurrent (index: Int) : Int{
+        return counters [index]
+    }
+    fun next (index: Int) : Int{
+        counters [index]++
+        return getCurrent(index)
+    }
+    fun set (index:Int, value : Int){
+        counters [index] = value
+    }
 }
