+++ app/src/main/java/net/svishch/android/githubclient/mvp/presenter/Presenter.kt
@@ -1,13 +1,13 @@
 package net.svishch.android.githubclient.mvp.presenter
 
-import net.svishch.android.githubclient.mvp.BtnIndexBase.Companion.btn1BaseIndex
-import net.svishch.android.githubclient.mvp.BtnIndexBase.Companion.btn2BaseIndex
-import net.svishch.android.githubclient.mvp.BtnIndexBase.Companion.btn3BaseIndex
-import net.svishch.android.githubclient.mvp.model.CountersModel
+import net.svishch.android.githubclient.mvp.presenter.BtnIndexBase.Companion.btn1BaseIndex
+import net.svishch.android.githubclient.mvp.presenter.BtnIndexBase.Companion.btn2BaseIndex
+import net.svishch.android.githubclient.mvp.presenter.BtnIndexBase.Companion.btn3BaseIndex
+import net.svishch.android.githubclient.mvp.model.Model
 import net.svishch.android.githubclient.mvp.view.MainView
 
 class Presenter(private val view: MainView){
-    private val model = CountersModel()
+    private val model = Model()
 
     fun counterClick(id: Int) {
         when (id) {
